⛑️ test : SubwayMap 단위 테스트 통합
